# files/swagger.yml
swagger: '2.0'
host: 'localhost:10001'
info:
  title: Gather
  description: Gather API Documentation
  version: 1.0.0
  contact:
    email: "mpreston@accedian.com"
basePath: "/api"
schemes:
  - http
consumes:
  - application/json
  - application/vnd.api+json
produces:
  - application/json
  - application/vnd.api+json
paths:
...
  '/v1/tenants/{tenantId}':
    get:
      summary: Retrieve a Tenant by id.
      operationId: GetTenant
      produces:
        - application/vnd.api+json
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/JsonApiTenant'
        '403':
          description: "Requestor does not have authorization to perform this action"
          schema:
            type: string
        '500':
          description: "Unexpected error processing request"
          schema:
            type: string
      parameters:
        - name: tenantId
          in: path
          required: true
          type: string
      tags:
        - AdminProvisioningService
    delete:
      summary: Delete a Tenant
      operationId: DeleteTenant
      produces:
        - application/vnd.api+json
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/JsonApiTenant'
        '403':
          description: "Requestor does not have authorization to perform this action"
          schema:
            type: string
        '500':
          description: "Unexpected error processing request"
          schema:
            type: string
      parameters:
        - name: tenantId
          in: path
          required: true
          type: string
      tags:
        - AdminProvisioningService
    patch:
      summary: Patch a Tenant
      operationId: PatchTenant
      consumes:
        - application/vnd.api+json
      produces:
        - application/vnd.api+json
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/JsonApiTenant'
        '400':
          description: "Request data does not pass validation"
        '403':
          description: "Requestor does not have authorization to perform this action"
          schema:
            type: string
        '500':
          description: "Unexpected error processing request"
          schema:
            type: string
      parameters:
        - name: tenantId
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/JsonApiTenant'
      tags:
        - AdminProvisioningService
...
definitions:
...
  JsonApiTenant:
    type: "object"
    properties:
      data:
        $ref: '#/definitions/Tenant'
...
  Tenant:
    type: "object"
    properties:
      type: 
        type: "string"
        enum: [tenants]
      id:
        type: "string"
      attributes:
        $ref: '#/definitions/TenantAttr'
  TenantAttr:
    type: "object"
    properties:
      _id:
        type: "string"
      _rev:
        type: "string"
      datatype:
        type: "string"
      name:
        type: "string"
      urlSubdomain:
        type: "string"
      state:
        type: "string"
        enum: [USER_UNKNOWN, INVITED, ACTIVE, SUSPENDED, PENDING_DELETE]
      createdTimestamp:
        type: "integer"
      lastModifiedTimestamp:
        type: "integer"
...